Module Web.HTTP

type headers

The type for header maps. A header map represents a list of HTTP headers. It maps header field names to their value.

type path = string list

The type for HTTP absolute_paths represented by its non-empty list of URI segments. Note that URI segments can be empty; in particular the absolute path "/" is the list with a single empty segment and is thus represented by the list [""].

type version = int * int

The type for representing HTTP-Version fields. Both integers must be in the interval [0;9].

type meth = [

| `GET

| `HEAD

| `POST

| `PUT

| `DELETE

| `CONNECT

| `OPTIONS

| `TRACE

| `PATCH

| `Other of string

]

The type for HTTP request methods.

module Headers : sig ... end

Headers. FIXME(dbuenzli): The semantic of this module considers than a field can contain multiple values. However, it's should be the wrong way to abstract values of headers - and let the user to consider than specific field can have multiple values.